+++ docs/source/changelog.rst
@@ -14,6 +14,8 @@ Dev branch
 * Tweak VMF ``localise()`` and ``translate()`` type hints to allow :py:class:`~srctools.math.FrozenVec` as the origin.
 * Make movement and rotation of displacements work correctly.
 * Handle pitch keyvalues correctly when instancing, only rotating if it is a specific type.
+* Changed :py:func:`srctools.instancing.collapse_one()` to use the entclass database directly,
+  deprecating the ``fgd`` parameter as a result.
 
 -------------
 Version 2.3.7